Fixed compiler errors

+++ include/model2.h
@@ -13,10 +13,10 @@
 class Model2
 {
     protected:
-        int warmup  = 30000;
-        int measure = 500000;
+        size_t warmup  = 30000;
+        size_t measure = 500000;
         static const int n_neigh = 4;
-        int size;
+        size_t size;
         bool isClean;
         std::uniform_real_distribution<float> rand0;
         std::vector<Neighbor<2>> neigh;
@@ -30,7 +30,7 @@ class Model2
         virtual double sweep_energy(double beta, std::mt19937 &engine) = 0;
         void set_exchange(double delta);
         std::vector<Exchange<2>> get_exchange() const;
-        void set_run_param(int Warmup, int Measure);
+        void set_run_param(size_t Warmup, size_t Measure);
 };
 
 #endif
